requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
datadog-api-client==1.6.0
idna==3.3
JSON-log-formatter==0.4.0
numpy==1.21.4
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
reverse-geocoder==1.5.1
scipy==1.7.3
six==1.16.0
urllib3==1.26.7
